% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestSumFunction.R
\name{sumTest}
\alias{sumTest}
\title{A Sum Test Function}
\usage{
sumTest(x, y)
}
\arguments{
\item{x}{first value to add together}

\item{y}{second value to add together}
}
\value{
x+y 
sumTest()
}
\description{
This function allows you to express your love of cats.
}
\keyword{sums}
